a few linear memory related improvements
* don't bother to defer allocation to acctual access for
  large linear memory. realloc overhead is more troublesome.

* lift an implementation limit and allow to use the full 32-bit
  linear memory.  (on 64-bit platforms)

* reduce the number of 64-bit arithmetics in the fast path.

+++ lib/type.h
@@ -378,7 +378,7 @@ struct meminst {
         uint8_t *data;
         /* Note: memory_getptr2 reads size_in_pages w/o locks */
         _Atomic uint32_t size_in_pages; /* overrides type->min */
-        uint32_t allocated;
+        size_t allocated;
         const struct memtype *type;
 
 #if defined(TOYWASM_ENABLE_WASM_THREADS)
